move get_stencils_with_varied_bounds to pace.dsl
