+++ fv3core/fv3core/decorators.py
@@ -2,14 +2,6 @@ import collections
 import collections.abc
 import functools
 import types
-from typing import Any, Callable, List, Mapping, Optional
-
-import fv3core
-import fv3core._config as spec
-import fv3core.utils
-import fv3core.utils.grid
-from pace.dsl.stencil import FrozenStencil, StencilFactory
-from pace.dsl.typing import Index3D
 
 
 ArgSpec = collections.namedtuple(
@@ -56,33 +48,3 @@ def get_namespace(arg_specs, state):
             namespace_kwargs[arg_name] = state[standard_name].storage
             namespace_kwargs[arg_name + "_quantity"] = state[standard_name]
     return types.SimpleNamespace(**namespace_kwargs)
-
-
-def get_stencils_with_varied_bounds(
-    func: Callable[..., None],
-    origins: List[Index3D],
-    domains: List[Index3D],
-    stencil_factory: StencilFactory,
-    externals: Optional[Mapping[str, Any]] = None,
-) -> List[FrozenStencil]:
-    assert len(origins) == len(domains), (
-        "Lists of origins and domains need to have the same length, you provided "
-        + str(len(origins))
-        + " origins and "
-        + str(len(domains))
-        + " domains"
-    )
-    if externals is None:
-        externals = {}
-    stencils = []
-    for origin, domain in zip(origins, domains):
-        ax_offsets = fv3core.utils.grid.axis_offsets(spec.grid, origin, domain)
-        stencils.append(
-            stencil_factory.from_origin_domain(
-                func,
-                origin=origin,
-                domain=domain,
-                externals={**externals, **ax_offsets},
-            )
-        )
-    return stencils
