+++ fv3core/tests/main/test_stencils_varied_bounds.py
@@ -1,8 +1,10 @@
 import numpy as np
 from gt4py.gtscript import PARALLEL, computation, horizontal, interval, region
 
-from fv3core.decorators import get_stencils_with_varied_bounds
-from pace.dsl.gt4py_utils import make_storage_from_shape
+from pace.dsl.gt4py_utils import (
+    get_stencils_with_varied_bounds,
+    make_storage_from_shape,
+)
 from pace.dsl.stencil import GridIndexing, StencilConfig, StencilFactory
 from pace.dsl.typing import FloatField
